Create README.md
My experience ranges from the IBM 1620 to the CDC 6000 series mainframes. Even a little Pascal on a wafer fab mainframe.

But mostly I’m a retired disk drive test engineer. As such I have programmed two IBM-style [mainframe channel adaptors] into standalone test boxes. Once I had to write a processor simulator for one of these processors. My code had to be 100% absolutely correct in how it handled tracks with identified flaws. 

Given my working standalone testing, I was asked to do something as directed by an IBM channel. I looked at what I was given and thought a bit. I asked that the command codes be changed into a this-bit-does-that code. It ended up saving our bacon because VM370 was a CPU hog doing I/O. I added a multi-track bit and an ending track to the command option applicable to all commands. The mainframe program now did 30 tracks as one I/O operation.

I also wrote a bit of CPU-implementing microcode. It was the only way to make selected decisions before the disk platter got to where the decision was needed. 

In the early days of a given product, I had to beg for a drive that would spin up and stay up. I would load drive code from the [then common ATA] interface. When writing waveforms the “scrambler” and other stuff had to be disabled. I used a scope to validate that the waveforms were written correctly.

For what was envisioned as a customer mode test program I made it reasonable to produce scripts that executed random script sequences. I made it reasonable to test the same script sequences in each of the 10 or more ATA forms.

My first programming was in absolute address, numeric opcode, and data fields. Then assembly language which was simply a non-mysterious better way to put stuff into a computer’s memory. 

I was directed to read Knuth when an application I was responsible for started printing several hundred error messages always about the same track. The consultants I was replacing used a bad LSFR random generator for a random seek test. This sparked an interest in the generation of mostly pseudo-random numbers.

My first submission to GitHub, pbRand, is my take on a multi-word PRNG. It comes with multiple seeding options as well as several ways to write generated randomness into memory. Four seed forms and three implantations as C++, x64 assembly, and 32-bit ARM assembly.